# apps/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Couture Dynamic Pro — API métier (Phase 4)
  version: "1.0"
  description: >
    Contrat des endpoints REST versionnés (collections, catalogues, pages,
    produits, variantes, panier, commandes). Les endpoints /api/auth et
    /api/admin/users (Phase 3) et les paiements (Phase 7) ne sont pas
    couverts par ce document.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cdp_session
...
paths:
  /collections:
    get:
      summary: Lister les collections (publiées pour le public, toutes pour ADMIN/CONTENT_MANAGER)
      security: []
      responses:
        "200":
          description: OK
    post:
      summary: Créer une collection (statut initial DRAFT)
      description: Rôles autorisés — ADMIN, CONTENT_MANAGER.
      responses:
        "201":
          description: Créée
        "400":
          description: Entrée invalide
        "401":
          description: Non authentifié
        "403":
          description: Rôle non autorisé
  /collections/{id}:
    get:
      summary: Détail d'une collection (404 si non publiée et appelant non staff)
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
        "404": { description: Non trouvée / non publiée }
    patch:
      summary: Mettre à jour une collection (y compris transition de statut)
      description: Rôles autorisés — ADMIN, CONTENT_MANAGER.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
        "404": { description: Non trouvée }
  /catalogues:
    get:
      summary: Lister les catalogues (publiés pour le public, tous pour le staff)
      security: []
      responses:
        "200": { description: OK }
    post:
      summary: Créer un catalogue (statut initial DRAFT, slug généré depuis le titre si absent)
      description: Rôles autorisés — ADMIN, CONTENT_MANAGER. RG4 (slug unique).
      responses:
        "201": { description: Créé }
        "409": { description: Slug déjà utilisé }
  /catalogues/{slug}:
    get:
      summary: Détail d'un catalogue par slug, avec ses pages (RG4)
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
        "404": { description: Non trouvé / non publié }
  /catalogues/{id}:
    patch:
      summary: Mettre à jour un catalogue ; publier applique RG3
      description: >
        Rôles autorisés — ADMIN, CONTENT_MANAGER. Le slug ne peut plus être
        modifié une fois PUBLISHED (RG4). Passer à PUBLISHED exige
        couverture + description + au moins une page (RG3).
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
        "400": { description: "RG3 non satisfait (NOT_PUBLISHABLE)" }
        "409": { description: "Slug verrouillé (publié) ou déjà utilisé" }
  /catalogues/{id}/pages:
    post:
      summary: Ajouter une page à un catalogue
      description: Rôles autorisés — ADMIN, CONTENT_MANAGER.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "201": { description: Créée }
        "409": { description: Numéro de page déjà utilisé (contrainte unique) }
  /catalogues/{id}/pages/{pageId}:
    delete:
      summary: Retirer une page d'un catalogue
      description: Rôles autorisés — ADMIN, CONTENT_MANAGER.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: pageId
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Supprimée }
...